feat: extract principal module
